Record.test.js => Record.test.ts
