+++ .gitignore
@@ -34,6 +34,7 @@
 /TypesUnit/_entity/JSONML.test.js
 /TypesUnit/_entity/Model.test.js
 /TypesUnit/_entity/ReactiveObject.test.js
+/TypesUnit/_entity/Record.test.js
 /TypesUnit/_entity/Time.test.js
 /TypesUnit/_formatter/**/*.js
 /TypesUnit/_function/**/*.js
